+++ CHANGELOG.md
@@ -6,6 +6,9 @@ a projekt používá [sémantické verzování](https://semver.org/lang/cs/).
 
 ## [Unreleased]
 
+### Fixed
+- Do výběru motivu se nově dostane jen sada s kompletními assety. Registr odmítne složku `cards_NN`, které chybí kterýkoli z 37 souborů (32 líců, rub, 4 ikony barev) nebo některé z obou pozadí, i složku s nečíselným/jednociferným označením. Dřív stačila libovolná složka `cards_*` s párem pozadí a neúplný motiv pak v náhledu i hře sypal chybějící obrázky (404).
+
 ## [2.2.0] - 2026-06-27
 
 ### Added